fix(generator): map unsigned long to uint64, not uint32
unsigned long is 64-bit on all LP64 platforms. Truncating to uint32
dropped the high half. Hoist ulong/long C-type preservation above the
goType switch to emit correct C.ulong casts without width mismatch.

Regenerates AVIOContext.checksum accessors: uint32 → uint64.

+++ structs.gen.go
@@ -14684,13 +14684,13 @@ func (s *AVIOContext) SetMinPacketSize(value int) {
 }
 
 // Checksum gets the checksum field.
-func (s *AVIOContext) Checksum() uint32 {
+func (s *AVIOContext) Checksum() uint64 {
 	value := s.ptr.checksum
-	return uint32(value)
+	return uint64(value)
 }
 
 // SetChecksum sets the checksum field.
-func (s *AVIOContext) SetChecksum(value uint32) {
+func (s *AVIOContext) SetChecksum(value uint64) {
 	s.ptr.checksum = (C.ulong)(value)
 }
